test(wasm): extract poison-recovery test to ratchet-exempt mod_tests.rs
The inline #[cfg(test)] poison-recovery block plus the struct doc pushed
api/mod.rs to 960 lines, over its 879 module-size ratchet budget (the
ratchet lives in ifc-lite-processing and scans repo-wide, so the
wasm-bindings crate test alone didn't catch it).

Move the test verbatim into a sibling `api/mod_tests.rs` (a `*_tests.rs`
file the ratchet exempts), declared `#[cfg(test)] mod mod_tests;`. As a
child of the `api` module it keeps `super` private-field access to
IfcAPI, so the test is unchanged. The full heap-safety reasoning moves
into that file's module doc; mod.rs keeps a one-line pointer. Net result
is 878 lines, so ratchet the allowlist budget down 879 -> 878.
