+++ rust/processing/tests/module_size_allowlist.txt
@@ -76,6 +76,6 @@
      767 rust/wasm-bindings/src/api/gpu_meshes/batch.rs
      727 rust/wasm-bindings/src/api/gpu_meshes/prepass.rs
      436 rust/wasm-bindings/src/api/grid_lines.rs
-     879 rust/wasm-bindings/src/api/mod.rs
+     878 rust/wasm-bindings/src/api/mod.rs
      692 rust/wasm-bindings/src/zero_copy/mesh.rs
      743 rust/wasm-bindings/src/zero_copy/symbolic.rs
