+++ rust/wasm-bindings/src/api/mod.rs
@@ -35,30 +35,6 @@ use wasm_bindgen::prelude::*;
 const TESSELLATION_QUALITY_MEDIUM: u8 = 2;
 
 /// Main IFC-Lite API
-///
-/// ## Poisoned-mutex recovery
-///
-/// Every `.lock()` call on the `Mutex` fields below recovers via
-/// `.unwrap_or_else(std::sync::PoisonError::into_inner)` instead of
-/// `.expect(...)`-panicking on a poisoned lock, so one malformed file no
-/// longer bricks a long-lived / multi-tenant `IfcAPI` instance (e.g. a
-/// parser worker reused across many `parse` calls) for the rest of its
-/// lifetime. This is safe for the cache slots (`cached_entity_index`,
-/// `cached_item_dedup`, `cached_parts_to_skip`, and friends): each is read
-/// and then replaced wholesale (`*slot = <freshly built value>` /
-/// `slot.take()`), never mutated field-by-field, so a panic while the lock
-/// is *held* necessarily happens before that one assignment runs — the
-/// guarded value is never torn, only ever the last known-good value or its
-/// untouched initial default. `pipeline_diagnostics` is the one exception:
-/// `record_pipeline_batch` mutates its counters in place
-/// (`ifc_lite_processing::PipelineDiagnostics::record_batch`), so a panic
-/// mid-call could in principle leave a partially-updated batch folded in.
-/// That accumulator is best-effort observability (surfaced only via
-/// `getPipelineDiagnostics`), never read for control flow or geometry
-/// correctness, and it keeps accumulating coherently from the next batch
-/// onward — recovering there trades a possible one-batch undercount in a
-/// diagnostics counter for not permanently losing the whole diagnostics
-/// channel to one poisoned lock.
 #[wasm_bindgen]
 pub struct IfcAPI {
     initialized: bool,
@@ -278,18 +254,7 @@ impl IfcAPI {
     /// the same `IfcAPI` instance — e.g. the parser worker keeps one
     /// `IfcAPI` alive across multiple `parse` requests).
     ///
-    /// Recovers from a poisoned cache Mutex instead of panicking
-    /// (`.lock().unwrap_or_else(PoisonError::into_inner)` everywhere these
-    /// cache mutexes are touched, see the module-level note on
-    /// `cached_entity_index`): every cache slot here is a whole-value
-    /// `Option<Arc<_>>` (or similar) that is only ever replaced after the
-    /// new value is fully built, so a poisoned lock means an earlier panic
-    /// happened elsewhere while the lock was merely *held*, not mid-write —
-    /// the guarded value itself was never torn. Recovering and clearing it
-    /// (this method's whole job) is therefore safe, and it turns "one
-    /// malformed file poisons the mutex and bricks every subsequent call on
-    /// a long-lived / multi-tenant `IfcAPI` instance" into "the next load
-    /// clears the cache and proceeds normally".
+    /// Recovers a poisoned cache Mutex instead of panicking; see `mod_tests.rs`.
     #[wasm_bindgen(js_name = clearPrePassCache)]
     pub fn clear_pre_pass_cache(&self) {
         let mut slot = self
@@ -910,51 +875,4 @@ fn set_js_prop(obj: &JsValue, key: &str, value: &JsValue) -> bool {
     js_sys::Reflect::set(obj, &JsValue::from_str(key), value).unwrap_or(false)
 }
 
-#[cfg(test)]
-mod poison_recovery_tests {
-    use super::IfcAPI;
-
-    /// A panic on another thread while it holds one of the cache mutexes
-    /// (e.g. a malformed entity deep in a lazy cache rebuild) must not
-    /// brick every later call on this `IfcAPI` instance. Poison
-    /// `cached_entity_index` directly, exactly like a real panicking
-    /// rebuild would, then assert `clearPrePassCache` — which locks every
-    /// cache mutex in turn — still returns normally instead of panicking
-    /// on `.expect("... Mutex poisoned")` (the pre-fix behaviour).
-    #[test]
-    fn clear_pre_pass_cache_recovers_from_a_poisoned_cache_mutex() {
-        let api = IfcAPI::new();
-
-        // Poison cached_entity_index: hold the lock on a spawned thread and
-        // panic while holding it. Unwinding through the MutexGuard's Drop
-        // is what marks a std::sync::Mutex poisoned.
-        let join_result = std::thread::scope(|scope| {
-            scope
-                .spawn(|| {
-                    let _guard = api.cached_entity_index.lock().unwrap();
-                    panic!("intentional poison for the poison-recovery test");
-                })
-                .join()
-        });
-        assert!(
-            join_result.is_err(),
-            "the spawned thread must have panicked while holding the lock"
-        );
-        assert!(
-            api.cached_entity_index.is_poisoned(),
-            "the mutex must be poisoned after the spawned thread panicked while holding it"
-        );
-
-        // Pre-fix, this call panicked on
-        // `.expect("ifc-lite cached_entity_index Mutex poisoned")`. It must
-        // now recover and clear every cache slot instead.
-        api.clear_pre_pass_cache();
-
-        // Recovering via `unwrap_or_else(PoisonError::into_inner)` does not
-        // clear the mutex's poison flag, so every future `.lock()` on it
-        // still observes `Err(PoisonError)` — the fix is recovering from
-        // that on *every* access, not a one-shot clear. Prove a second call
-        // also succeeds instead of panicking again.
-        api.clear_pre_pass_cache();
-    }
-}
+#[cfg(test)] mod mod_tests;
